requirements.txt
dash==2.7.0
dash-pivottable==0.0.2
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
requests==2.28.1
python-dotenv==0.21.0